Fix import of charms to env storage
